+++ src/main/java/me/anisekai/server/events/interest/InterestLevelUpdatedEvent.java
@@ -3,9 +3,9 @@ package me.anisekai.server.events.interest;
 import me.anisekai.server.entities.Interest;
 import me.anisekai.server.events.InterestUpdatedEvent;
 
-public class InterestLevelUpdatedEvent extends InterestUpdatedEvent<Long> {
+public class InterestLevelUpdatedEvent extends InterestUpdatedEvent<Byte> {
 
-    public InterestLevelUpdatedEvent(Object source, Interest entity, Long previous, Long current) {
+    public InterestLevelUpdatedEvent(Object source, Interest entity, Byte previous, Byte current) {
 
         super(source, entity, previous, current);
     }
